GithubCommand: Defer all initial replies
We need to respond to Discord's chat interaction request within 3
seconds. If we don't do so, we sometimes end up in a perpetual broken
state where chat commands simply stop working.

This changes GithubCommand to always send out an ephemeral initial
deferred reply, which then either gets deleted and replaced with the
actual reply or edited when something goes wrong. We need to delete the
initial deferred reply, since the ephemeral status of a message cannot
be changed with `.editReply()`.