GithubCommand: Edit initial reply if successful
Making the initial deferred reply ephemeral caused the follow-up
messages to show an ugly "Message could not be loaded" at the top, since
the ephemeral message was deleted. And even if we didn't delete it,
other users would not be able to load it.

This changes the initial reply to be visible to everyone and get edited
on success. On failure, the initial reply is deleted again and a
follow-up ephemeral message is posted to the user.